refactor: move create_intensity_component to physics
