+++ tests/test_model.py
@@ -1,16 +1,10 @@
 # pylint: disable=redefined-outer-name
-
 import numpy as np
 import pytest
 import sympy as sp
-from expertsystem.amplitude.helicity import HelicityModel
 
 from tensorwaves.interfaces import DataSample, Function
-from tensorwaves.model import (
-    LambdifiedFunction,
-    SympyModel,
-    create_intensity_component,
-)
+from tensorwaves.model import LambdifiedFunction, SympyModel
 
 
 @pytest.fixture(scope="module")
@@ -52,19 +46,6 @@ def test_complex_amplitude(
     np.testing.assert_array_almost_equal(results, expected_results, decimal=4)
 
 
-def test_helicity(helicity_model: SympyModel):
-    assert set(helicity_model.parameters) == {
-        "C[J/\\psi(1S) \\to f_{0}(980)_{0} \\gamma_{+1};f_{0}(980) \\to \\pi^{0}_{0} \\pi^{0}_{0}]",
-        "C[J/\\psi(1S) \\to f_{0}(500)_{0} \\gamma_{+1};f_{0}(500) \\to \\pi^{0}_{0} \\pi^{0}_{0}]",
-        "m_f(0)(980)",
-        "d_f(0)(980)",
-        "Gamma_f(0)(980)",
-        "m_f(0)(500)",
-        "d_f(0)(500)",
-        "Gamma_f(0)(500)",
-    }
-
-
 def test_canonical(canonical_model: SympyModel):
     assert set(canonical_model.parameters) == {
         "C[J/\\psi(1S) \\to f_{0}(980)_{0} \\gamma_{+1};f_{0}(980) \\to \\pi^{0}_{0} \\pi^{0}_{0}]",
@@ -78,32 +59,14 @@ def test_canonical(canonical_model: SympyModel):
     }
 
 
-def test_create_intensity_component(
-    phsp_set: DataSample,
-    es_helicity_model: HelicityModel,
-    intensity: LambdifiedFunction,
-):
-    # pylint: disable=line-too-long
-    model = es_helicity_model
-    from_amplitudes = create_intensity_component(
-        model,
-        components=[
-            R"A[J/\psi(1S)_{+1} \to f_{0}(500)_{0} \gamma_{+1};f_{0}(500)_{0} \to \pi^{0}_{0} \pi^{0}_{0}]",
-            R"A[J/\psi(1S)_{+1} \to f_{0}(980)_{0} \gamma_{+1};f_{0}(980)_{0} \to \pi^{0}_{0} \pi^{0}_{0}]",
-        ],
-        backend="numpy",
-    )
-    from_intensity = create_intensity_component(
-        model,
-        components=R"I[J/\psi(1S)_{+1} \to \gamma_{+1} \pi^{0}_{0} \pi^{0}_{0}]",
-        backend="numpy",
-    )
-    assert pytest.approx(from_amplitudes(phsp_set)) == from_intensity(phsp_set)
-
-    intensity_components = [
-        create_intensity_component(model, component, backend="numpy")
-        for component in model.components
-        if component.startswith("I")
-    ]
-    sub_intensities = [i(phsp_set) for i in intensity_components]
-    assert pytest.approx(sum(sub_intensities)) == intensity(phsp_set)
+def test_helicity(helicity_model: SympyModel):
+    assert set(helicity_model.parameters) == {
+        "C[J/\\psi(1S) \\to f_{0}(980)_{0} \\gamma_{+1};f_{0}(980) \\to \\pi^{0}_{0} \\pi^{0}_{0}]",
+        "C[J/\\psi(1S) \\to f_{0}(500)_{0} \\gamma_{+1};f_{0}(500) \\to \\pi^{0}_{0} \\pi^{0}_{0}]",
+        "m_f(0)(980)",
+        "d_f(0)(980)",
+        "Gamma_f(0)(980)",
+        "m_f(0)(500)",
+        "d_f(0)(500)",
+        "Gamma_f(0)(500)",
+    }
